Add search to Rails Guides
For better discoverability and to prevent users from having to leave
the Guides webiste, this commit implements a client side search based
on the open lunr library.

Different libraries were compared:

* lucaong/minisearch, where it was not possible to pre-build an index.
  Building it on the fly would take about 3 seconds and render the
  browser unresponsive.
* krisk/Fuse, where it was possible to pre-build an index, but the
  fuzzy search was very slow and would block for about 600ms.
* olivernn/lunr.js, where the index cloud be pre-built and the fuzzy
  search could be disabled.

The new search uses Lunr with a pre-built index and by using a lookup
table where the id of the documents are mapped to their respective url
to keep the index size small.

To build the index, the Guides HTML is parsed and split into documents
that are then indexed using ExecJS. This generates an index of about
3.3 MB in size which is only loaded when the user focuses the search
input.

The index is then stored in the browsers localstorage for up to a week
for performance.